feature refactorizado codigo
